Moving indirect funding components to protocols directory
