+++ packages/wallet/src/redux/protocols/indirect-funding/components/approve-funding.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import ApproveX from '../approve-x';
+import ApproveX from '../../../../components/approve-x';
 import web3Utils from 'web3-utils';
 
 interface Props {
