+++ packages/wallet/src/redux/protocols/indirect-funding/player-a/container.tsx
@@ -1,7 +1,7 @@
 import React, { PureComponent } from 'react';
 import { connect } from 'react-redux';
-import ApproveFunding from '../../../../components/indirect-funding/approve-funding';
-import { FundingStep } from '../../../../components/indirect-funding/funding-step';
+import ApproveFunding from '../components/approve-funding';
+import { FundingStep } from '../components/funding-step';
 import * as actions from './actions';
 import * as indirectFundingPlayerA from './state';
 import { unreachable } from '../../../../utils/reducer-utils';
